fix: move pluginmain to pkg
