+++ plugins/cmd/litellm/main.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"time"
 
-	"github.com/naira-project/naira/plugins/internal/pluginmain"
+	"github.com/naira-project/naira/plugins/pkg/pluginmain"
 )
 
 type config struct {
