+++ plugins/pkg/pluginmain/pluginmain.go
